fix(s2): address review feedback (reconnect lifecycle, naming, docs, tests)
