+++ crates/adapters/src/transport/s2/mod.rs
@@ -5,5 +5,5 @@ mod test;
 
 pub use input::S2InputEndpoint;
 #[cfg(test)]
-pub(crate) use input::{Metadata as S2Metadata, make_replay_read_input};
+pub(crate) use input::{S2CheckpointMetadata as S2Metadata, make_replay_read_input};
 pub use output::S2OutputEndpoint;
