+++ docs.feldera.com/docs/connectors/sinks/s2.md
@@ -16,12 +16,16 @@ with the `s2_output` connector.
 | `auth_token` | string | Yes | S2 authentication token. Use a [secret reference](/connectors/secret-references) in production. |
 | `endpoint` | string | No | Custom S2 endpoint, primarily for local development. When omitted, the S2 cloud endpoint is used. |
 
-The connector submits each encoded output buffer as one S2 record. S2 limits a
-record body to 1 MiB, so the connector asks the encoder to keep each buffer below
-that limit. The connector waits for S2 to acknowledge the append before returning.
-It is not a fault-tolerant output connector; a failure after S2 accepted a record
-but before Feldera observed its acknowledgement can produce a duplicate after a
-retry.
+The connector splits each encoded output buffer into one S2 record per line and
+appends them as native S2 batches, chunked to S2's per-batch record limit. S2
+enforces a 1 MiB metered limit per `AppendRecordBatch` (including per-record
+framing overhead); this is a hard, server-side limit, not a Feldera setting. The
+connector asks the encoder to keep each buffer below that limit, but a single
+oversized record — for example, a row with a very long string column — can still
+exceed it and will be rejected by the append call. The connector waits for S2 to
+acknowledge each batch before returning. It is not a fault-tolerant output
+connector; a failure after S2 accepted a batch but before Feldera observed its
+acknowledgement can produce duplicates after a retry.
 
 ## Example
