+++ docs.feldera.com/docs/connectors/sources/s2.md
@@ -21,12 +21,19 @@ checkpointed ranges when a pipeline resumes.
 `start_from` accepts the following values:
 
 - `"Beginning"`: start at sequence number 0.
-- `"Tail"`: consume records appended after the connector starts.
+- `"Tail"`: consume records appended after the connector starts. The tail is
+  resolved once, right before the first record is read, and that sequence number
+  is then checkpointed, so it does not keep drifting as the stream grows. Before
+  the first record, the anchor is time-dependent; if you need a fully
+  deterministic start, use `SeqNum` or `Timestamp`.
 - `{"SeqNum": 42}`: start at sequence number 42.
 - `{"Timestamp": 1773900000000}`: start at a Unix timestamp in milliseconds.
 - `{"TailOffset": 100}`: start 100 records before the current tail.
 
 Once a checkpoint exists, its sequence number takes precedence over `start_from`.
+If the checkpoint was taken after the start position was resolved (a record was
+consumed or a tail was anchored), the connector always resumes from that absolute
+sequence number — including sequence 0 — rather than recomputing `start_from`.
 The connector preserves ordered delivery across reconnects and validates sequence
 contiguity after the start position has been resolved. Gaps, duplicates, or S2
 trimmed records required to replay a checkpoint fail the connector rather than
